Complete ChangeButton component
I think this is all we need for the change button. It will trigger a
re-render on each click for the whole form and change its text from
'Edit' to 'Submit' based on the editing status passed to it as a prop.
I don't think the change button needs any more functionality itself
at this time.